[PD] Change bootstrap_room metadata dtype from int64 to uint64 (#19141)

+++ python/sglang/srt/disaggregation/utils.py
@@ -167,7 +167,7 @@ class MetadataBuffers:
             )
             # Request validation: store bootstrap_room to detect metadata corruption
             self.bootstrap_room = torch.zeros(
-                (size, 8), dtype=bootstrap_room_dtype, device=device
+                (size, 8), dtype=torch.uint64, device=device
             )
 
     def get_buf_infos(self):
